requirements.txt
absl-py==2.1.0
ace-tools @ file:///home/sandbox/.openai_internal/ace-tools
aeppl==0.0.31
aesara==2.7.3
affine==2.4.0
aiohttp==3.8.6
aiosignal==1.3.1
analytics-python==1.4.post1
annotated-types==0.7.0
anyio==3.7.1
anytree==2.8.0
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
arviz==0.19.0
asn1crypto==1.5.1
asttokens==2.4.1
async-timeout==4.0.3
attrs==24.2.0
audioread==3.0.1
babel==2.16.0
backoff==1.10.0
basemap==1.3.9
basemap-data==1.3.2
bcrypt==4.2.0
beautifulsoup4==4.12.3
bleach==6.1.0
blinker==1.8.2
blis==0.7.11
blosc2==2.0.0
bokeh==2.4.0
branca==0.7.2
Brotli==1.1.0
cachetools==5.5.0
cairocffi==1.7.1
CairoSVG==2.5.2
camelot-py==0.10.1
catalogue==2.0.10
certifi==2024.7.4
cffi==1.17.0
chardet==3.0.4
charset-normalizer==2.1.1
click==8.1.7
click-plugins==1.1.1
cligj==0.7.2
cloudpathlib==0.18.1
cloudpickle==3.0.0
cmake==3.30.2
cmudict==1.0.28
comm==0.2.2
confection==0.1.5
cons==0.4.6
contourpy==1.2.1
countryinfo==0.1.2
cryptography==3.4.8
cssselect2==0.7.0
cycler==0.12.1
cymem==2.0.8
Cython==0.29.36
databricks-sql-connector==0.9.1
debugpy==1.8.5
decorator==4.4.2
defusedxml==0.7.1
dlib==19.24.2
dm-tree==0.1.8
dnspython==2.6.1
docx2txt==0.8
einops==0.3.2
email_validator==2.2.0
entrypoints==0.4
et-xmlfile==1.1.0
etuples==0.3.9
exchange-calendars==3.4
executing==2.0.1
Faker==8.13.2
fastapi==0.111.0
fastapi-cli==0.0.5
fastjsonschema==2.20.0
fastprogress==1.0.3
ffmpeg-python==0.2.0
ffmpy==0.4.0
filelock==3.15.4
Fiona==1.9.2
Flask==3.0.3
Flask-CacheBuster==1.0.0
Flask-Cors==4.0.1
Flask-Login==0.6.3
folium==0.12.1
fonttools==4.53.1
fpdf==1.7.2
frozenlist==1.4.1
future==1.0.0
fuzzywuzzy==0.18.0
gensim==4.3.1
geographiclib==1.52
geopandas==0.10.2
geopy==2.2.0
gradio==2.2.15
graphviz==0.17
gTTS==2.2.3
h11==0.14.0
h2==4.1.0
h5netcdf==1.3.0
h5py==3.8.0
hpack==4.0.0
html5lib==1.1
httpcore==1.0.5
httptools==0.6.1
httpx==0.27.0
hypercorn==0.14.3
hyperframe==6.0.1
idna==3.7
imageio==2.35.1
imageio-ffmpeg==0.5.1
imgkit==1.2.2
importlib_metadata==8.4.0
importlib_resources==6.4.3
iniconfig==2.0.0
ipykernel==6.29.5
ipython==8.26.0
ipython-genutils==0.2.0
isodate==0.6.1
itsdangerous==2.2.0
jax==0.2.28
jedi==0.19.1
Jinja2==3.1.4
joblib==1.4.2
json5==0.9.25
jsonpickle==3.2.2
jsonschema==4.23.0
jsonschema-specifications==2023.12.1
jupyter-server==1.23.5
jupyter_client==7.4.9
jupyter_core==5.1.3
jupyterlab==3.4.8
jupyterlab-pygments==0.2.2
jupyterlab_server==2.19.0
keras==2.6.0
kerykeion==2.1.16
kiwisolver==1.4.5
korean-lunar-calendar==0.3.1
langcodes==3.4.0
language_data==1.2.0
lazy_loader==0.4
librosa==0.8.1
lit==18.1.8
llvmlite==0.43.0
logical-unification==0.4.6
loguru==0.5.3
lxml==5.3.0
marisa-trie==1.2.0
markdown-it-py==3.0.0
markdown2==2.5.0
markdownify==0.9.3
MarkupSafe==2.1.5
matplotlib==3.6.3
matplotlib-inline==0.1.7
matplotlib-venn==0.11.6
mdurl==0.1.2
miniKanren==1.0.3
mistune==3.0.2
mizani==0.10.0
mne==0.23.4
monotonic==1.6
moviepy==1.0.3
mpmath==1.3.0
msgpack==1.0.8
mtcnn==0.1.1
multidict==6.0.5
multipledispatch==1.0.0
munch==4.0.0
murmurhash==1.0.10
mutagen==1.45.1
nashpy==0.0.35
nbclassic==0.4.5
nbclient==0.10.0
nbconvert==7.16.4
nbformat==5.10.4
nest-asyncio==1.6.0
networkx==2.8.8
nltk==3.6.3
notebook==6.5.1
notebook_shim==0.2.4
numba==0.60.0
numexpr==2.10.1
numpy==1.24.0
numpy-financial==1.0.0
nvidia-cublas-cu11==11.10.3.66
nvidia-cuda-cupti-cu11==11.7.101
nvidia-cuda-nvrtc-cu11==11.7.99
nvidia-cuda-runtime-cu11==11.7.99
nvidia-cudnn-cu11==8.5.0.96
nvidia-cufft-cu11==10.9.0.58
nvidia-curand-cu11==10.2.10.91
nvidia-cusolver-cu11==11.4.0.1
nvidia-cusparse-cu11==11.7.4.91
nvidia-nccl-cu11==2.14.3
nvidia-nvtx-cu11==11.7.91
odfpy==1.4.1
opencv-python==4.5.5.62
openpyxl==3.0.10
opt-einsum==3.3.0
orjson==3.10.7
oscrypto==1.3.0
packaging==24.1
pandas==1.5.3
pandocfilters==1.5.1
paramiko==3.4.1
parso==0.8.4
patsy==0.5.6
pdf2image==1.16.3
pdfkit==0.6.1
pdfminer.six==20220319
pdfplumber==0.6.2
pdfrw==0.4
pexpect==4.9.0
Pillow==9.2.0
platformdirs==4.2.2
plotly==5.3.0
plotnine==0.10.1
pluggy==1.5.0
pooch==1.8.2
preshed==3.0.9
priority==2.0.0
proglog==0.1.10
prometheus_client==0.20.0
prompt_toolkit==3.0.47
pronouncing==0.2.0
psutil==6.0.0
ptyprocess==0.7.0
pure_eval==0.2.3
py==1.11.0
py-cpuinfo==9.0.0
pycountry==20.7.3
pycparser==2.22
pycryptodome==3.20.0
pycryptodomex==3.20.0
pydantic==2.7.2
pydantic-extra-types==2.9.0
pydantic-settings==2.4.0
pydantic_core==2.18.3
pydot==1.4.2
pydub==0.25.1
pydyf==0.11.0
Pygments==2.18.0
pygraphviz==1.7
PyJWT==2.9.0
pylog==1.1
pyluach==2.2.0
pymc==4.0.1
PyMuPDF==1.21.1
PyNaCl==1.5.0
pyOpenSSL==21.0.0
pypandoc==1.6.3
pyparsing==3.1.2
PyPDF2==1.28.6
pyphen==0.16.0
pyproj==3.6.1
pyprover==0.5.6
pyshp==2.3.1
pyswisseph==2.10.3.2
pytesseract==0.3.8
pytest==6.2.5
pyth3==0.7
python-dateutil==2.9.0.post0
python-docx==0.8.11
python-dotenv==1.0.1
python-json-logger==2.0.7
python-multipart==0.0.9
python-pptx==0.6.21
pyttsx3==2.90
pytz==2024.1
PyWavelets==1.7.0
pyxlsb==1.0.8
PyYAML==6.0.2
pyzbar==0.1.8
pyzmq==26.1.1
qrcode==7.3
rarfile==4.0
rasterio==1.3.3
rdflib==6.0.0
referencing==0.35.1
regex==2024.7.24
reportlab==3.6.12
requests==2.31.0
resampy==0.4.3
rich==13.7.1
rpds-py==0.20.0
scikit-image==0.20.0
scikit-learn==1.1.3
scipy==1.9.3
seaborn==0.11.2
Send2Trash==1.8.3
sentencepiece==0.2.0
shap==0.39.0
Shapely==1.7.1
shellingham==1.5.4
six==1.16.0
slicer==0.0.7
smart-open==7.0.4
sniffio==1.3.1
snowflake-connector-python==2.7.12
snuggs==1.4.7
SoundFile==0.10.2
soupsieve==2.6
spacy==3.7.5
spacy-legacy==3.0.12
spacy-loggers==1.0.5
srsly==2.4.8
stack-data==0.6.3
starlette==0.37.2
statsmodels==0.13.5
svglib==1.1.0
svgwrite==1.4.1
sympy==1.8
tables==3.8.0
tabula==1.0.5
tabulate==0.9.0
tenacity==9.0.0
terminado==0.18.1
text-unidecode==1.3
textblob==0.15.3
thinc==8.2.5
threadpoolctl==3.5.0
thrift==0.20.0
tifffile==2024.8.10
tinycss2==1.3.0
toml==0.10.2
tomli==2.0.1
toolz==0.12.1
torch==2.0.1
torchaudio==2.0.2
torchtext==0.6.0
torchvision==0.15.2
tornado==6.4.1
tqdm==4.64.0
traitlets==5.14.3
trimesh==3.9.29
triton==2.0.0
typer==0.12.4
typing_extensions==4.10.0
ujson==5.10.0
urllib3==1.26.19
uvicorn==0.19.0
uvloop==0.20.0
Wand==0.6.13
wasabi==1.1.3
watchfiles==0.23.0
wcwidth==0.2.13
weasel==0.4.1
weasyprint==53.3
webencodings==0.5.1
websocket-client==1.8.0
websockets==10.3
Werkzeug==3.0.3
wordcloud==1.9.2
wrapt==1.16.0
wsproto==1.2.0
xarray==2024.3.0
xarray-einstats==0.7.0
xgboost==1.4.2
XlsxWriter==3.2.0
xml-python==0.4.3
yarl==1.9.4
zipp==3.20.0
zopfli==0.2.3
